+++ packages/poi/lib/utils/AppError.js
@@ -1,6 +1,6 @@
 module.exports = class AppError extends Error {
   constructor(msg) {
     super(msg)
-    this.name = this.constructor.name
+    this.name = 'AppError'
   }
 }
