+++ packages/poi/lib/utils/arrify.js
@@ -1,0 +1,5 @@
+function arrify(input) {
+  return Array.isArray(input) ? input : [input]
+}
+
+module.exports = arrify
