+++ packages/poi/lib/utils/createSet.js
@@ -1,0 +1,5 @@
+function createSet(value) {
+  return Array.isArray(value) ? new Set(value) : new Set([value])
+}
+
+module.exports = createSet
