+++ packages/poi/lib/utils/dir.js
@@ -1,0 +1,9 @@
+const path = require('path')
+
+exports.cwd = function(cwd, ...args) {
+  return path.resolve(cwd || process.cwd(), ...args)
+}
+
+exports.ownDir = function(...args) {
+  return path.join(__dirname, '../../', ...args)
+}
