+++ packages/poi/lib/utils/getFullEnvString.js
@@ -1,0 +1,8 @@
+function getFullEnvString(env) {
+  return Object.keys(env).reduce((res, key) => {
+    res[`process.env.${key}`] = env[key]
+    return res
+  }, {})
+}
+
+module.exports = getFullEnvString
