+++ packages/poi/lib/utils/getPublicPath.js
@@ -1,0 +1,8 @@
+function getPublicPath(command, homepage) {
+  if (command === 'build' && typeof homepage === 'string') {
+    return /\/$/.test(homepage) || homepage === '' ? homepage : homepage + '/'
+  }
+  return '/'
+}
+
+module.exports = getPublicPath
