+++ packages/poi/lib/utils/isPath.js
@@ -1,0 +1,5 @@
+function isPath(v) {
+  return /.+\..+/.test(v) || v.indexOf('/') > -1
+}
+
+module.exports = isPath
