+++ packages/poi/lib/utils/unspecifiedAddress.js
@@ -1,0 +1,5 @@
+function unspecifiedAddress(host) {
+  return host === '0.0.0.0' || host === '::'
+}
+
+module.exports = unspecifiedAddress
