+++ packages/poi/lib/webpack/imageRules.js
